requirements.txt
neo4j>=5.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0

# Veri toplama (web scraping + LLM extraction)
scrapling[fetchers]>=0.4.0
anthropic>=0.40.0
